requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.1; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
async-timeout==4.0.3; python_version < '3.11'
attrs==24.2.0; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.2; python_version < '3.11'
fastapi==0.112.2; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.8; python_version >= '3.6'
multidict==6.0.5; python_version >= '3.7'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pymupdf==1.23.7; python_version >= '3.8'
pymupdfb==1.23.7; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.38.2; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
uvicorn==0.30.6; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
